Moved validation of tree and application relation to TreeEventHandler
