+++ src/main/java/org/sitmun/domain/tree/Tree.java
@@ -12,7 +12,6 @@ import org.sitmun.domain.tree.node.TreeNode;
 import org.sitmun.domain.user.User;
 import org.sitmun.infrastructure.persistence.type.codelist.CodeList;
 import org.sitmun.infrastructure.persistence.type.i18n.I18n;
-import org.sitmun.infrastructure.persistence.type.tree.TreeType;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
@@ -30,7 +29,6 @@ import java.util.Set;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-@TreeType
 public class Tree {
 
   /**
